feat(title): Move Tux out from behind the main menu
Tux and the menu both sat mid-screen, so he walked the whole demo behind it.
The view is now pushed along so he passes beside it. Other menus will cover
him, but at least now we get to see the penguin walk when on the main menu.

Tux also used to teleport back to the start from plain sight at the end of the
level. Now Tux walks off the right and then teleports to the left once off screen.
He then arrives in from the left side walking starting the loop again. The first
arrival still falls into shot as it always did. It's just later laps he walks in.

Moving the view took whole numbers, so it couldn't be held still against Tux
sliding along by fractions and it'd judder as it stopped. We fix this so it now
takes fractions.

+++ src/object/camera.hpp
@@ -69,7 +69,7 @@ public:
    * to the position goal
    */
   void scroll_to(const Vector& goal, float scrolltime);
-  void move(const int dx, const int dy);
+  void move(const float dx, const float dy);
 
   void reload_config();
